+++ src/components/molecules/DownloadDropdown/DownloadDropdown.styled.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const DownloadWrapper = styled.div`
+  display: flex;
+  position: relative;
+  justify-content: center;
+  align-items: center;
+  gap: 0.5rem;
+`;
