+++ src/components/organisms/Header/Header.tsx
@@ -1,24 +1,15 @@
 import Finder from '../../../assets/img/Finder.svg';
-import { useContext, useState } from 'react';
+import { useContext } from 'react';
 import { ThemeContext } from '../../../theme/ThemeContext';
 import Typography from '../../atoms/Typography/Typography';
 import * as Styled from './Header.styled';
-import Icon from '../../atoms/Icon/Icon';
 import Button from '../../atoms/Button/Button';
 import SearchBar from '../../molecules/SearchBar/SearchBar';
-import QRDropdownItem from '../../molecules/QRDropdownItem/QRDropdownItem';
 import NotificationDropdown from '../../molecules/NotificationDropdown/NotificationDropdown';
+import DownloadDropdown from '../../molecules/DownloadDropdown/DownloadDropdown';
 
 const Header = () => {
   const { palette } = useContext(ThemeContext);
-  const [isQRDropdownOpen, setIsQRDropdownOpen] = useState<boolean>(false);
-
-  const openQRDropdown = () => {
-    setIsQRDropdownOpen(true);
-  };
-  const closeQRDropdown = () => {
-    setIsQRDropdownOpen(false);
-  };
 
   return (
     <Styled.HeaderWrapper palette={palette}>
@@ -28,13 +19,8 @@ const Header = () => {
           <Styled.Logo src={Finder} alt='Logo Finder' />
         </Styled.LogoAndHambuergerContainer>
         <Styled.SearchAndButtonsContainer>
-          <Styled.DownloadWrapper onMouseOver={openQRDropdown} onMouseOut={closeQRDropdown}>
-            <Icon name='download' />
-            <Typography color='blueDark' variant='boldParagraph'>
-              App
-            </Typography>
-            {isQRDropdownOpen ? <QRDropdownItem /> : null}
-          </Styled.DownloadWrapper>
+          
+          <DownloadDropdown />
 
           <SearchBar placeholder='Search' />
